tests/wpt/web-platform-tests/websockets/Create-wrong-scheme.htm

<!DOCTYPE html>
<html>
<head>
    <title>W3C WebSocket API - Create WebSocket - non ws/wss scheme in url</title>
    <script type="text/javascript" src="/resources/testharness.js"></script>
    <script type="text/javascript" src="/resources/testharnessreport.js"></script>
    <script type="text/javascript" src="websocket.sub.js"></script>
</head>
<body>
    <div id="log"></div>
    <script type="text/javascript">
        test(function () {
            var wsocket;
            assert_throws("SYNTAX_ERR", function () { wsocket = CreateWebSocketNonWsScheme() });
        }, "W3C WebSocket API - Create WebSocket - Pass a URL with a non ws/wss scheme - SYNTAX_ERR is thrown")
    </script>
</body>
</html>
